+++ src/com/neurix/simrs/transaksi/antrianonline/model/RegistrasiOnline.java
@@ -3,6 +3,7 @@ package com.neurix.simrs.transaksi.antrianonline.model;
 import com.neurix.common.model.BaseModel;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.sql.Date;
 import java.sql.Timestamp;
 
@@ -15,7 +16,7 @@ public class RegistrasiOnline extends BaseModel implements Serializable {
     private String noKtp;
     private String tempatLahir;
     private Date tglLahir;
-    private Long desaId;
+    private BigInteger desaId;
     private String jalan;
     private String suku;
     private String agama;
@@ -108,11 +109,11 @@ public class RegistrasiOnline extends BaseModel implements Serializable {
         this.tglLahir = tglLahir;
     }
 
-    public Long getDesaId() {
+    public BigInteger getDesaId() {
         return desaId;
     }
 
-    public void setDesaId(Long desaId) {
+    public void setDesaId(BigInteger desaId) {
         this.desaId = desaId;
     }
